requirements.txt
python==3.11.8
numpy==1.26.4
pytorch==2.0.1
torchvision==0.15.2
mamba-ssm==2.0.4
scipy==1.5.4
h5py==3.1.0
addict==2.4.0